+++ tools/libexec/perl/TTP/HTTP/Compare/Config.pm
@@ -83,6 +83,7 @@ use constant {
 		"\.xls\$"
 	],
 	DEFAULT_CRAWL_BY_CLICK_INTERMEDIATE_SCREENSHOTS => false,
+	DEFAULT_CRAWL_BY_CLICK_RESTORE_CHAIN_LAST => 10,
 	DEFAULT_CRAWL_BY_CLICK_TEXT_DENY_PATTERNS => [
 	],
 	DEFAULT_CRAWL_BY_CLICK_XPATH_DENY_PATTERNS => [
@@ -774,6 +775,22 @@ sub confCrawlByClickIntermediateScreenshots {
 	$enabled = DEFAULT_CRAWL_BY_CLICK_INTERMEDIATE_SCREENSHOTS if !defined $enabled;
 
 	return $enabled;
+}
+
+# ------------------------------------------------------------------------------------------------
+# Returns the configured count of falsy restore chain which cancel the role
+# (I):
+# - none
+# (O):
+# - returns the configured count of falsy restore chain
+
+sub confCrawlByClickRestoreChainLast {
+	my ( $self ) = @_;
+
+	my $count = $self->var([ 'crawl', 'by_click', 'restore_chain_last' ]);
+	$count = DEFAULT_CRAWL_BY_CLICK_RESTORE_CHAIN_LAST if !defined $count;
+
+	return $count;
 }
 
 # ------------------------------------------------------------------------------------------------